+++ src/prodockit/__init__.py
@@ -101,6 +101,6 @@ alongside them.
 See https://prodockit.org/ for documentation.
 """
 
-__version__ = "0.43.0a1"
+__version__ = "0.43.0a2"
 
 __all__ = ["__version__"]